API UPDATE - See Notes
API /AddPersonToElevator was updated to send back HTTP code 400 if the personID or elevatorID don't exist.
API /AddPersonToElevator was updated to check if the elevator can visit the floors the user will want.
API /AddPersonToElevator_A3 was updated to send back HTTP code 400 if the personID or elevatorID don't exist.
API /AddPersonToElevator_A3 was updated to check if the elevator can visit the floors the user will want.