API Corrections
The API was corrected for the AddPersonToElevator and AddPersonToElevator_A3 endpoints to now protect the system from being told to add someone to an elevator if they have already been assigned.  This should prevent the ability for people to clone themselves when bad input is passed to the system.  If the person has been handled previously the system will now return HTTP code 400 and the error message: "Person ID {personID} was already assigned previously."

Beyond that, Tomas' update for the ElevatorStatus endpoint was duplicated down to the ElevatorStatus_A3 API.